refactor : 레포지토리 -> 도메인으로 디렉구조 변경 #24
