+++ src/main/kotlin/com/get_offer/product/domain/ProductRepository.kt
@@ -1,7 +1,5 @@
-package com.get_offer.product.repository
+package com.get_offer.product.domain
 
-import com.get_offer.product.domain.Product
-import com.get_offer.product.domain.ProductStatus
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.jpa.repository.JpaRepository
